move internal implementation details to a separate header
